Added the Tester functional interface.
public boolean isEnabled();

intially, the Tester's functional method was named test() because you can
use it to perform any custom test that you want.

However, test() hides the primary nature and purpose of this interface.

So it was renamed to isEnabled().

The primary purpose is to check if the plugin isEnabled() BEFORE
isCompatible() or isSupported() is performed...

Because, if the plugin is not even running on the server, then any
compatibility checks should fail (as intended).

public boolean isEnabled();

return true if you want isCompatible() & isSupported to proceed.
return false if you want them to stop and also return false.

The Version constructor that doesn't ask you to pass in a Tester will
constructor a null tester that always returns true.