Added generics
New optional syntax:
Version<Plugin> p = VersionFactory.getPluginVersion("SomePlugin");
Version<Server> s = VersionFactory.getServerVersion();

Also, I really dislike the name VersionFactory. It should probably be
changed to BukkitVersionFactory, since it is specific to Bukkit plugins
and servers. Altho, I also don't like how long it is. Plus, this is a
breaking change and would require all dependent code to be updated. So i'm
undecided.

All of this is probably overly complicated just because I wanted to
provide a little extra functionality built-in: And that is checking if a
plugin is even enabled along with the compatibility checks.