Deprecated, doc'd, future-proofed getNmsVersion().
Problem:
The whole problem is that versioned packaging may not exist in future
versions of Spigot. Also, it does not exist in really old versions of
Craftbukkit (older than 1.4.5).

So, this update is a safe-guard against possible future changes in Spigot
while maintaining backwards compatibility. In other words, getNmsVersion()
will check for the existence of net.minecraft.server.vX_Y_RZ. Previously,
if this failed, we simply assumed that it's an old version and we returned
the value "pre"... meaning that it's prior to versions packaging.

Another problem is the return value. The main goal of this method is to
reliably detect the server version and construct a String "vX_Y_Z" that
can be used as a package name to access:

com.name.project.nms.vX_Y_Z;
com.name.project.compat.vX_Y_Z;

Therefore, the real return value should be String, NOT a Version object.

Deprecated:
getNmsVersion in favor or private getNmsPackage().

Future updates v2.0:
getNmsVersion() will be removed.
getNmsPackage() will be public.
getNmsPackage() will no longer return "pre"

Documented:
The javadocs @return now document that getNmsVersion() returns "pre" for
servers less than 1.4.5.