Refactored Version.java
to use the new Tester container...

because ultimately, we only care about tester.test() and if it succeeds or
fails.

Whereas, previously, we were asking for both a Predicate tester and the
Object testee... And sometimes, the testee might actually be null...
wierd.

So now, all of these details are abstracted away and encapsulated into the
Tester class, whose job is to succeed or fail, and relay this information
to the Version object's compatibility methods().

Basically, Tester is an optional parameter for the Version constructor,
which allows additional tests() to be executed before compatibility checks
are performed: For example, doing plugin.isEnabled() ?

Wait, the plugin isn't enabled, you say ?

Okay, great. No need to continue any further. We can stop right here,
because any plugin that is disabled is NOT compatible.

Which reminds me: That for a plugin's onEnable() method, an Exception can
occur, which might render some of its fields null. In which case, the
Exception wouldn't disable the plugin... it would still be active and
enabled... but with null fields, there's a good likelihood that the
plugin will be broken. And in this case, maybe compatibility check should
also fail ? To prevent problems from occuring due to the interaction with
a broken plugin.