Adds FieldTester
The problem is that server admins were installing broken versions of
WorldEdit. Where an exception would occur during startup, which meant that
not all the fields were initialized.

So, checking that WorldEdit is installed and checking the version is NOT
enough: We also have to check to make sure that WorldEdit was properly
initialized (all their fields are not null).

This was the main use-case scenario for FieldTester.

All examples can be found in FieldTesterTest.java and
TesterFactoryTest.java